removing external reserve()
