+++ contracts/libs/ERC20Reservable.sol
@@ -7,7 +7,7 @@ pragma solidity ^0.8.13;
 
 import '@openzeppelin/contracts-upgradeable/token/ERC20/ERC20Upgradeable.sol';
 
-contract Reservable is ERC20Upgradeable {
+contract ERC20Reservable is ERC20Upgradeable {
     enum ReservationStatus {
         Draft, // 0
         Active, // 1
@@ -52,19 +52,6 @@ contract Reservable is ERC20Upgradeable {
             ReservationStatus.Active
         );
         _totalReserved[from_] += amount_ + executionFee_;
-    }
-
-    function reserve(
-        address from_,
-        address to_,
-        address executor_,
-        uint256 amount_,
-        uint256 executionFee_,
-        uint256 nonce_,
-        uint256 deadline_
-    ) external returns (bool success) {
-        _reserve(from_, to_, executor_, amount_, executionFee_, nonce_, deadline_);
-        return true;
     }
 
     function reserveOf(address account_) external view returns (uint256 count) {
