+++ package.json
@@ -24,7 +24,6 @@
     "ethereum-waffle": "^3.0.0",
     "ethers": "5.0.0",
     "hardhat": "^2.9.3",
-    "hardhat-awesome-cli": "^0.0.11",
     "hardhat-gas-reporter": "^1.0.8",
     "pre-commit": "^1.2.2",
     "prettier": "^2.6.2",
@@ -32,7 +31,9 @@
     "solidity-coverage": "^0.7.21"
   },
   "dependencies": {
-    "@openzeppelin/contracts-upgradeable": "^4.6.0"
+    "@openzeppelin/contracts-upgradeable": "^4.6.0",
+    "dotenv": "^16.0.1",
+    "hardhat-awesome-cli": "^0.0.15"
   },
   "pre-commit": [
     "prettier-test",
