+++ test/shared.js
@@ -3,10 +3,10 @@ const { ethers, network, addressBook } = require('hardhat');
 const { expect, use } = require('chai');
 const { solidity } = require('ethereum-waffle');
 use(solidity);
-const DECIMALS = 18;
 
 const NAME = 'IP3';
 const SYMBOL = 'IP3';
+const DECIMALS = 18;
 const TOTALSUPPLY = ethers.utils.parseUnits('300000000000', DECIMALS);
 const VERSION = '0.1';
 const VERSION_712 = '1';
